# OPT: recursive subdataset discovery can be made faster