--- a/datalad/distribution/subdatasets.py
+++ b/datalad/distribution/subdatasets.py
@@ -173,6 +173,11 @@ class Subdatasets(Interface):
             dataset, check_installed=False, purpose='subdataset reporting')
         refds_path = dataset.path
 
+        # XXX this seems strange, but is tested to be the case -- I'd rather set
+        # `check_installed` to true above and fail
+        if not GitRepo.is_valid_repo(refds_path):
+            return
+
         # return as quickly as possible
         if isinstance(recursion_limit, int) and (recursion_limit <= 0):
             return