````
% time datalad subdatasets -r
datalad subdatasets -r  1,48s user 0,93s system 20% cpu 11,775 total
````

reference:

````
% time git submodule status --recursive
git submodule status --recursive  0,20s user 0,13s system 11% cpu 2,784 total
````

approach seems simple.