@mih 2017-05-09
nipype  example fails.