@mih 2017-05-10
This is not true. Dont trust those coverage reports.



On May 10, 2017 19:23, "Yaroslav Halchenko" <notifications@github.com>
wrote:

> *@yarikoptic* commented on this pull request.
>
> I started to write "I wonder if someone could come up..." and ended up
> with #1513 <https://github.com/datalad/datalad/pull/1513>
> ------------------------------
>
> In datalad/distribution/subdatasets.py
> <https://github.com/datalad/datalad/pull/1507#discussion_r115738299>:
>
> >      """All known ones with some properties"""
> -    modinfo = _parse_gitmodules(dspath)
> +    # this will not work in direct mode, need better way #1422
> +    cmd = ['git', '--work-tree=.', 'submodule', 'status']
> +    if recursive:
> +        cmd.append('--recursive')
>
> wow -- we don't have a test testing for recursive invocation of this
> function anywhere??
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/datalad/datalad/pull/1507#pullrequestreview-37301577>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAIVH7Uj_9HglW4XNTpwQSR9iPaiRIGwks5r4fKVgaJpZM4NVUlL>
> .
>