@yarikoptic 2017-05-10
well -- example fails with
```shell
[INFO   ] aggregating meta data for <Dataset path=/tmp/datalad_demo_nipype.7m9K/datasets.datalad.org/workshops/nipype-2017> 

[INFO   ] aggregating meta data for <Dataset path=/tmp/datalad_demo_nipype.7m9K/datasets.datalad.org/workshops> 

save(notneeded): /tmp/datalad_demo_nipype.7m9K/datasets.datalad.org/workshops/nipype-2017/ds000114 (dataset)

save(ok): /tmp/datalad_demo_nipype.7m9K/datasets.datalad.org/workshops/nipype-2017 (dataset)

[ERROR  ] Failed to run ['git', '-c', 'receive.autogc=0', '-c', 'gc.auto=0', 'add', '--dry-run', '-N', '--ignore-missing', '--verbose', 'nipype-2017/ds000114', 'nipype-2017', '.'] under '/tmp/datalad_demo_nipype.7m9K/datasets.datalad.org/workshops'. Exit code=128. out= err=fatal: Pathspec 'nipype-2017/ds000114' is in submodule 'nipype-2017'
|  
fatal: Pathspec 'nipype-2017/ds000114' is in submodule 'nipype-2017'

I: done running docs/examples/nipype_workshop_dataset.sh: failed
```
so while dealing with submodules and this PR digs into the holy grail of their handling seems to me... doesn't failure reproduce if you run that example locally with your changes?