@mih 2017-05-10
Yes, it fails -- reliably. But you can easily convince yourself that we have tests that call recursive subdataset queries.