@yarikoptic 2017-05-10
since I trust noone (even myself when convinced) -- added
```
     if recursive:
+        raise RuntimeError("Yarik was wrong")
         cmd.append('--recursive')
```
and was really convinced... but then WTF coverage report is wrong???? bleh  I guess will investigate whenever all tests pass -- may be somehow caused by a failure