@mih 2017-05-12
Failure in example is due to #1518 (a bug that is in master already, but only becomes visible due to the changes in here).